JsonNodeMarshallUnmarshallContextTesting2 extracted JsonNodeMarshallUnmarshallContextTesting
